Fix greatest increase in profits calculation
Updates getGreatestIncrease() to use the rate of changes returned by getRateOfChanges().

The logic will find the greatest RoC from all the periods, and finds the correspondent initial date of that period.

When finding the present transaction of the RoC need to add 1 to the index of the max RoC found.